pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.6",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "plotly>=6.1.2",
    "pytesseract>=0.3.13",
    "scikit-learn>=1.7.0",
    "streamlit>=1.45.1",
]
